+++ sklearn/datasets/base.py
@@ -20,56 +20,12 @@ from os.path import isdir
 from os.path import splitext
 from os import listdir
 from os import makedirs
+from ..utils import Bunch
 
 import numpy as np
 
 from ..utils import check_random_state
 
-
-class Bunch(dict):
-    """Container object for datasets
-
-    Dictionary-like object that exposes its keys as attributes.
-
-    >>> b = Bunch(a=1, b=2)
-    >>> b['b']
-    2
-    >>> b.b
-    2
-    >>> b.a = 3
-    >>> b['a']
-    3
-    >>> b.c = 6
-    >>> b['c']
-    6
-
-    """
-
-    def __init__(self, **kwargs):
-        super(Bunch, self).__init__(kwargs)
-
-    def __setattr__(self, key, value):
-        self[key] = value
-
-    def __dir__(self):
-        return self.keys()
-
-    def __getattr__(self, key):
-        try:
-            return self[key]
-        except KeyError:
-            raise AttributeError(key)
-
-    def __setstate__(self, state):
-        # Bunch pickles generated with scikit-learn 0.16.* have an non
-        # empty __dict__. This causes a surprising behaviour when
-        # loading these pickles scikit-learn 0.17: reading bunch.key
-        # uses __dict__ but assigning to bunch.key use __setattr__ and
-        # only changes bunch['key']. More details can be found at:
-        # https://github.com/scikit-learn/scikit-learn/issues/6196.
-        # Overriding __setstate__ to be a noop has the effect of
-        # ignoring the pickled __dict__
-        pass
 
 
 def get_data_home(data_home=None):
