+++ sklearn/pipeline.py
@@ -20,7 +20,7 @@ from .externals.joblib import Parallel, delayed, Memory
 from .externals import six
 from .utils import tosequence
 from .utils.metaestimators import if_delegate_has_method
-from .datasets.base import Bunch
+from .utils import Bunch
 
 __all__ = ['Pipeline', 'FeatureUnion']
 
@@ -123,7 +123,7 @@ class Pipeline(_BasePipeline):
 
     Attributes
     ----------
-    named_steps : bunch object
+    named_steps : bunch object, a dictionary with attribute access
         Read-only attribute to access any step parameter by user given name.
         Keys are step names and values are steps parameters.
